refactor: move chat page fetch API to custom hooks
