+++ src/components/conversation/pages/ChatPage.jsx
@@ -1,116 +1,28 @@
 import { Bot, CircleUser, Send } from 'lucide-react';
 import React, { useEffect, useRef, useState } from 'react';
 import { useParams } from 'react-router';
+import useGetConversation from '../../../hooks/useGetConversation';
+import useSendMessage from '../../../hooks/useSendMessage';
 
 export default function ChatPage() {
     const params = useParams();
-    const [messages, setMessages] = useState([]);
     const [newMessage, setNewMessage] = useState('');
     const messagesEndRef = useRef(null);
 
-    useEffect(() => {
-        async function getConversation() {
-            try {
-                const response = await fetch(`http://localhost:4000/v1/messages/${params.conversationId}`);
-                const data = await response.json();
-                setMessages(data.messages || []);
-            } catch (error) {
-                console.error('Error fetching messages:', error);
-            }
-        }
-        getConversation();
-    }, [params.conversationId]);
+    const { data: messages, loading, error, setMessages } = useGetConversation(params.conversationId);
+    const { handleSubmit } = useSendMessage(params.conversationId, setMessages, setNewMessage);
 
     useEffect(() => {
         messagesEndRef.current?.scrollIntoView({ behavior: 'smooth' });
     }, [messages]);
 
-    async function handleSubmit(event) {
-        event.preventDefault();
-        if (!newMessage.trim()) return;
-
-        const userMessage = {
-            id: new Date().toISOString(),
-            type: 'user',
-            message: newMessage,
-            createdTime: new Date().toISOString(),
-        };
-        setMessages((prev) => [...prev, userMessage]);
-        setNewMessage('');
-
-        try {
-            const response = await fetch(`http://localhost:4000/v1/create-message`, {
-                method: 'POST',
-                headers: {
-                    'Content-Type': 'application/json',
-                },
-                body: JSON.stringify({
-                    userQuery: newMessage,
-                    conversationId: params.conversationId,
-                }),
-            });
-
-            if (!response.body) throw new Error('No response body');
-
-            const reader = response.body.getReader();
-            const decoder = new TextDecoder('utf-8');
-            let buffer = '';
-            let assistantContent = '';
-            const tempAssistantId = `assistant-${Date.now()}`;
-
-            setMessages((prev) => [
-                ...prev,
-                {
-                    id: tempAssistantId,
-                    type: 'assistant',
-                    message: '',
-                    createdTime: new Date().toISOString(),
-                }
-            ]);
-
-            while (true) {
-                const { done, value } = await reader.read();
-                if (done) break;
-
-                buffer += decoder.decode(value, { stream: true });
-
-                const parts = buffer.split('\n\n');
-                buffer = parts.pop();
-
-                for (let part of parts) {
-                    if (part.startsWith('data: ')) {
-                        const jsonString = part.replace(/^data: /, '');
-                        const parsed = JSON.parse(jsonString);
-
-                        if (parsed.state === 'response') {
-                            assistantContent += parsed.content || '';
-
-                            setMessages((prev) =>
-                                prev.map((msg) =>
-                                    msg.id === tempAssistantId
-                                        ? { ...msg, message: assistantContent }
-                                        : msg
-                                )
-                            );
-                        } else if (parsed.state === 'complete') {
-                            return;
-                        }
-                    }
-                }
-            }
-        } catch (err) {
-            console.error('Streaming error:', err);
-        }
-    }
-
-
     return (
         <div className='w-[60%] flex flex-col h-[90%] self-center'>
             <div className="flex-1 overflow-y-auto mb-4 space-y-4 p-4 border-0 rounded-lg no-scrollbar">
-                {messages.length === 0 ? (
+                {messages?.length === 0 ? (
                     <p className="text-gray-500 text-center">No messages to display</p>
                 ) : (
-                    messages.map((message) => (
+                    messages?.map((message) => (
                         <div key={message.id} className="flex flex-col hover:bg-stone-700 border-0 rounded-md p-4">
                             <div className={`flex gap-2 text-sm items-center ${message.type === "user" ? 'text-rose-500' : 'text-blue-400'}`}>
                                 {message.type === 'user' ? <CircleUser /> : <Bot />}
@@ -125,7 +37,7 @@ export default function ChatPage() {
                 <div ref={messagesEndRef} />
             </div>
 
-            <form onSubmit={handleSubmit} className='flex justify-center items-center w-full bg-stone-800 gap-4 focus-within:ring-1 border-0 rounded-full text-sm focus-within:ring-stone-600'>
+            <form onSubmit={(e) => handleSubmit(e, newMessage)} className='flex justify-center items-center w-full bg-stone-800 gap-4 focus-within:ring-1 border-0 rounded-full text-sm focus-within:ring-stone-600'>
                 <input
                     className="w-[90%] h-9 bg-stone-800 focus:outline-none"
                     type="text"
